+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## future
 ### New features
 - Add message to remove songs from library with `remove-from-library`.
+- Add option to set custom config path.
 
 ### GUI
 - Add hovers.